# models/models.gen.go
// Package models provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.12.4 DO NOT EDIT.
package models

// Address defines model for Address.
type Address = string

// AddressFilter defines model for AddressFilter.
type AddressFilter = Address

// Age defines model for Age.
type Age = int64

// CountResponse defines model for CountResponse.
type CountResponse struct {
	Total int64 `json:"total"`
}

// Dag defines model for Dag.
type Dag struct {
	Age              Age   `json:"age"`
	Hash             Hash  `json:"hash"`
	Level            int64 `json:"level"`
	TransactionCount int64 `json:"transactionCount"`
}

// DagsPaginatedResponse defines model for DagsPaginatedResponse.
type DagsPaginatedResponse = PaginatedResponse

// Hash defines model for Hash.
type Hash = string

// PaginatedResponse defines model for PaginatedResponse.
type PaginatedResponse struct {
	End     Hash  `json:"end"`
	HasNext bool  `json:"hasNext"`
	Start   Hash  `json:"start"`
	Total   int64 `json:"total"`
}

// PaginationFilter defines model for PaginationFilter.
type PaginationFilter struct {
	Limit *int32 `json:"limit,omitempty"`
	Start *Hash  `json:"start,omitempty"`
}

// Pbft defines model for Pbft.
type Pbft struct {
	Age              Age   `json:"age"`
	Hash             Hash  `json:"hash"`
	Number           int64 `json:"number"`
	TransactionCount int64 `json:"transactionCount"`
}

// PbftsPaginatedResponse defines model for PbftsPaginatedResponse.
type PbftsPaginatedResponse = PaginatedResponse

// Transaction defines model for Transaction.
type Transaction struct {
	Age      Age     `json:"age"`
	Block    int64   `json:"block"`
	From     Address `json:"from"`
	GasPrice string  `json:"gasPrice"`
	GasUsed  string  `json:"gasUsed"`
	Hash     Hash    `json:"hash"`
	Index    int64   `json:"index"`
	Status   bool    `json:"status"`
	To       Address `json:"to"`
	Value    string  `json:"value"`
}

// TransactionsPaginatedResponse defines model for TransactionsPaginatedResponse.
type TransactionsPaginatedResponse = PaginatedResponse

// Validator defines model for Validator.
type Validator struct {
	Address   Address `json:"address"`
	PbftCount int64   `json:"pbftCount"`
}

// ValidatorsPaginatedResponse defines model for ValidatorsPaginatedResponse.
type ValidatorsPaginatedResponse = PaginatedResponse

// WeekFilter defines model for WeekFilter.
type WeekFilter struct {
	Week int32 `json:"week"`
	Year int32 `json:"year"`
}

// GetAddressDagsParams defines parameters for GetAddressDags.
type GetAddressDagsParams struct {
	// Pagination Pagination
	Pagination *PaginationFilter `form:"pagination,omitempty" json:"pagination,omitempty"`
}

// GetAddressPbftsParams defines parameters for GetAddressPbfts.
type GetAddressPbftsParams struct {
	// Pagination Pagination
	Pagination *PaginationFilter `form:"pagination,omitempty" json:"pagination,omitempty"`
}

// GetAddressTransactionsParams defines parameters for GetAddressTransactions.
type GetAddressTransactionsParams struct {
	// Pagination Pagination
	Pagination *PaginationFilter `form:"pagination,omitempty" json:"pagination,omitempty"`
}

// GetValidatorsParams defines parameters for GetValidators.
type GetValidatorsParams struct {
	// Week Week to filter by
	Week WeekFilter `form:"week" json:"week"`

	// Pagination Pagination
	Pagination *PaginationFilter `form:"pagination,omitempty" json:"pagination,omitempty"`
}

// GetValidatorsTotalParams defines parameters for GetValidatorsTotal.
type GetValidatorsTotalParams struct {
	// Filter Week to filter by
	Filter WeekFilter `form:"filter" json:"filter"`
}